+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/Weapon.java
@@ -44,6 +44,7 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.npcs.MirrorImage;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.KindOfWeapon;
 import com.shatteredpixel.shatteredpixeldungeon.items.LiquidMetal;
+import com.shatteredpixel.shatteredpixeldungeon.items.MetalPart;
 import com.shatteredpixel.shatteredpixeldungeon.items.bags.Bag;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfArcana;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfForce;
@@ -309,9 +310,8 @@ abstract public class Weapon extends KindOfWeapon {
 	@Override
 	public ArrayList<String> actions(Hero hero) {
 		ArrayList<String> actions = super.actions(hero);
-		if (hero.hasTalent(Talent.APART_ANYTHING) && hero.heroClass != HeroClass.ENGINEER
-				&& !this.unique
-				&& !isEquipped(hero) && cursedKnown && !cursed && !hasCurseEnchant())
+		if (hero.pointsInTalent(Talent.APART_ANYTHING) >= 2
+				&& !this.unique && !isEquipped(hero) && cursedKnown && !cursed)
 			actions.add(AC_SMELT);
 		return actions;
 	}
@@ -319,14 +319,21 @@ abstract public class Weapon extends KindOfWeapon {
 	@Override
 	public void execute(Hero hero, String action) {
 		super.execute(hero, action);
-		if (action.equals(AC_SMELT)){
-			LiquidMetal metal = new LiquidMetal();
-			int quantity = (int)Math.pow(2, level()) * (tier + 1) * 3;
-			if (enchantment != null) quantity = Math.round(quantity * 1.5f);
-			if (hero.pointsInTalent(Talent.APART_ANYTHING) >= 2) quantity = Math.round(quantity * 1.67f);
+		if (action.equals(AC_SMELT) && hero.pointsInTalent(Talent.APART_ANYTHING) >= 2
+				&& !isEquipped(hero) && cursedKnown && !cursed){
+			if (hero.heroClass == HeroClass.ENGINEER) {
+				MetalPart metal = new MetalPart();
+				metal.quantity(level() + 1);
+				if (!metal.collect()) Dungeon.level.drop(metal, hero.pos).sprite.drop();
 
-			metal.quantity(quantity);
-			if (!metal.collect()) Dungeon.level.drop(metal, hero.pos).sprite.drop();
+			} else {
+				LiquidMetal metal = new LiquidMetal();
+				int quantity = (int) Math.pow(2, level()) * (tier + 1) * 3;
+				if (enchantment != null) quantity = Math.round(quantity * 1.5f);
+
+				metal.quantity(quantity);
+				if (!metal.collect()) Dungeon.level.drop(metal, hero.pos).sprite.drop();
+			}
 
 			detach(hero.belongings.backpack);
 			hero.sprite.operate(hero.pos);
